Builds Proxy Server Logic
Adds HTML page that combines the three services into one page. Includes
basic styling to get them fitted to the page.

Adds server logic that fetches the data for each of the services and
loads it into that component. This way each are updated based on which
room is loaded.